chore: simplify reload module logic
